using 32-bit sequence numbers
In real, using 31 bit sequence numbers, represented as a signed 32 bit.
The central data store will also be able to handle 32 bit signed
integers only.

+++ sensor/sensor.hpp
@@ -26,7 +26,7 @@ public:
 private:
 	void send();
 
-	uint16_t seq_ = 0;
+	int32_t seq_ = 0;
 	unsigned short magic_;
 	unsigned char id_;
 	aes128_ctx_t aes_ctx_;
